backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
bcrypt==4.1.2
python-socketio==5.10.0
eventlet==0.33.3
PyJWT==2.8.0
requests==2.31.0
python-docx==1.1.0
websocket-client==1.6.4

